requirements.txt
pyspellchecker==0.6.0
pyyaml